@@ -6,7 +6,7 @@ import "./App.css";
 import HomePage from "./pages/HomePage/HomePage";
 import LoginPage from "./pages/LoginPage/LoginPage";
 import RegisterPage from "./pages/RegisterPage/RegisterPage";
-
+import BlogPage from "./pages/BlogPage/Blog";
 // Component Imports
 import Navbar from "./components/NavBar/NavBar";
 import Footer from "./components/Footer/Footer";
@@ -27,8 +27,9 @@ function App() {
             </PrivateRoute>
           }
         />
-        <Route path="/register" element={<RegisterPage />} />
-        <Route path="/login" element={<LoginPage />} />
+        <Route path="/register" element={ <RegisterPage /> } />
+        <Route path="/login" element={ <LoginPage /> } />
+        <Route path="/blog" element={ <BlogPage /> } />
       </Routes>
       <Footer />
     </div>
