@@ -1,44 +1,100 @@
-import React from 'react';
-import 'bootstrap/dist/css/bootstrap.min.css';
+import React, { useState, useEffect } from 'react';
+import { useNavigate, Link } from 'react-router-dom';
+import axios from 'axios';
+import { Container, Row, Col, Card, Button, Alert } from 'react-bootstrap';
+import AuthContext from '../../context/AuthContext';
 
 const Blog = () => {
+  const { user, isAdmin } = React.useContext(AuthContext);
+  const navigate = useNavigate();
+  const [posts, setPosts] = useState([]);
+  const [error, setError] = useState('');
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchPosts = async () => {
+      try {
+        const response = await axios.get('http://127.0.0.1:8000/api/blog/');
+        setPosts(response.data);
+        setError('');
+      } catch (err) {
+        setError(err.response?.data?.detail || 'Failed to load blog posts');
+      } finally {
+        setLoading(false);
+      }
+    };
+    fetchPosts();
+  }, []);
+
+  const handleCreatePost = () => {
+    navigate('/blog/create');
+  };
+
   return (
-    <div className="container mt-5">
-      <div className="row">
-        <div className="col-md-10 offset-md-1">
-          <h1 className="mb-4">Analysis of US Patents for WiFi/LiFi Open-Source Edge Routers</h1>
-          <h2 className="mb-3">Key Points</h2>
-          <ul className="mb-4">
-            <li>Research suggests there are US patents for WiFi and LiFi technologies relevant to edge routers, but open-source LiFi routers face patent challenges.</li>
-            <li>It seems likely that open-source projects like Koruza offer alternatives, though they use infrared, not visible light LiFi.</li>
-            <li>The evidence leans toward a heavily patented LiFi field, requiring careful navigation for commercial home use.</li>
-          </ul>
-          <h2 className="mb-3">Patents and Technologies</h2>
-          <p>US patents exist for both WiFi and LiFi, covering router systems and light-based communication. For WiFi, patents like <a href="https://patents.google.com/patent/US7529230B2/en" target="_blank" rel="noopener noreferrer">US7529230B2</a> address routing functionality. For LiFi, patents such as <a href="https://patents.justia.com/patent/10992383" target="_blank" rel="noopener noreferrer">US10992383</a> and <a href="https://patents.google.com/patent/US20160218807A1/en" target="_blank" rel="noopener noreferrer">US20160218807A1</a> detail light-based networks, potentially applicable to home use.</p>
-          <p>Additionally, research indicates a broader patent filing trend in LiFi, with entities like Siemens, Samsung, Huawei, and LiFi Labs filing hundreds of applications in the US, EPO, CN, and WIPO, as noted in a 2017 article (<a href="https://legaladvantage.net/2017/06/li-fi-light-fidelity/" target="_blank" rel="noopener noreferrer">Legal Advantage</a>). This suggests a heavily patented field, which could impact open-source development.</p>
-          <h2 className="mb-3">Open-Source Considerations</h2>
-          <p>Open-source edge routers, like those using OpenWrt firmware for WiFi, are common, but hardware designs are less prevalent. For LiFi, the Koruza project (<a href="https://en.wikipedia.org/wiki/Koruza_%28technology%29" target="_blank" rel="noopener noreferrer">Koruza</a>) is an open-source, open-hardware initiative using infrared light, supporting high data rates but not specifically a router. LiFi's patented nature means open-source projects must avoid infringement, posing challenges for commercial deployment.</p>
-          <h2 className="mb-3">Implications for Commercial Home Use</h2>
-          <p>For commercial home use, both WiFi and LiFi technologies have patents that could apply, but LiFi's patent intensity poses unique challenges. WiFi patents, such as those for router designs and presence detection, are well-established and often licensed through standards bodies like IEEE, facilitating open-source firmware development. In contrast, LiFi's patents, like US10992383 and US20160218807A1, cover core technologies that could be essential for router functionality, potentially requiring licensing for commercial deployment.</p>
-          <p>The Koruza project offers a potential model for open-source optical communication, but its infrared focus may limit direct LiFi router applications. Given Houston's growing tech ecosystem, integrating LiFi into homes could leverage existing lighting infrastructure, but patent navigation will be critical. Legal strategies, such as partnering with patent holders or focusing on non-patented aspects, will be essential for your business.</p>
-          <h2 className="mb-3">Challenges and Opportunities</h2>
-          <p>The heavily patented LiFi field presents challenges for open-source development, as noted in articles discussing patent litigation, such as Huawei's disputes with Netgear over WiFi 6 patents (<a href="https://www.juve-patent.com/cases/huawei-and-bird-bird-successful-against-amazon-in-dispute-over-wifi-routers/" target="_blank" rel="noopener noreferrer">JUVE Patent</a>). This suggests potential legal risks for open-source LiFi routers, requiring thorough patent analysis. However, opportunities exist in niche markets, such as secure home networks, where LiFi's interference-free nature could be advantageous, as highlighted by the LiFi Group (<a href="https://lifi.co/" target="_blank" rel="noopener noreferrer">LiFi Group</a>).</p>
-          <h3 className="mt-5 mb-3">Key Citations</h3>
-          <ul>
-            <li><a href="https://patents.google.com/patent/US7529230B2/en" target="_blank" rel="noopener noreferrer">US7529230B2 - Wireless router system and method</a></li>
-            <li><a href="https://patents.google.com/patent/USD640692S1/en" target="_blank" rel="noopener noreferrer">USD640692S1 - Wireless router</a></li>
-            <li><a href="https://patents.google.com/patent/WO2017087285A1/en" target="_blank" rel="noopener noreferrer">WO2017087285A1 - System and method of wifi router based presence detection and control</a></li>
-            <li><a href="https://patents.justia.com/patent/10992383" target="_blank" rel="noopener noreferrer">US10992383 - Ad hoc light-based mesh network</a></li>
-            <li><a href="https://patents.google.com/patent/US20160218807A1/en" target="_blank" rel="noopener noreferrer">US20160218807A1 - Lifi communication system</a></li>
-            <li><a href="https://en.wikipedia.org/wiki/Koruza_%28technology%29" target="_blank" rel="noopener noreferrer">Koruza - Wikipedia</a></li>
-            <li><a href="https://legaladvantage.net/2017/06/li-fi-light-fidelity/" target="_blank" rel="noopener noreferrer">Legal Advantage - Li-Fi (Light Fidelity)</a></li>
-            <li><a href="https://opensource.com/article/22/7/openwrt-open-source-firmware" target="_blank" rel="noopener noreferrer">OpenWrt, an open source alternative to firmware for home routers</a></li>
-            <li><a href="https://www.juve-patent.com/cases/huawei-and-bird-bird-successful-against-amazon-in-dispute-over-wifi-routers/" target="_blank" rel="noopener noreferrer">Huawei and Bird & Bird successful against Amazon in dispute over wifi routers</a></li>
-            <li><a href="https://lifi.co/" target="_blank" rel="noopener noreferrer">LiFi Group - Revolutionary Wireless Communication Technology</a></li>
-          </ul>
-        </div>
-      </div>
-    </div>
+    <Container className="my-5">
+      <Row className="mb-4">
+        <Col>
+          <h1>LiFi Blog: Empowering Houston with Light-Based Connectivity</h1>
+          <p>
+            Discover how LiFi technology is revolutionizing internet access in Houston’s
+            healthcare, energy, and logistics sectors with unmatched speed and security.
+          </p>
+          {isAdmin() && (
+            <Button variant="primary" onClick={handleCreatePost}>
+              Create New Post
+            </Button>
+          )}
+        </Col>
+      </Row>
+
+      {error && (
+        <Alert variant="danger" className="mb-4">
+          {error}
+        </Alert>
+      )}
+
+      {loading ? (
+        <p>Loading posts...</p>
+      ) : posts.length === 0 ? (
+        <p>No posts available. Check back soon!</p>
+      ) : (
+        <Row xs={1} md={2} lg={3} className="g-4">
+          {posts.map((post) => (
+            <Col key={post.id}>
+              <Card>
+                <Card.Body>
+                  <Card.Title>{post.title}</Card.Title>
+                  <Card.Text>
+                    {post.is_premium && !user ? (
+                      <span>
+                        Premium content.{' '}
+                        <Link to="/login">Log in</Link> to read more.
+                      </span>
+                    ) : (
+                      post.content.substring(0, 150) + '...'
+                    )}
+                  </Card.Text>
+                  <Card.Text>
+                    <small className="text-muted">
+                      By {post.author} on{' '}
+                      {new Date(post.created_at).toLocaleDateString()}
+                    </small>
+                  </Card.Text>
+                  <Button
+                    variant="outline-primary"
+                    as={Link}
+                    to={`/blog/${post.id}`}
+                    disabled={post.is_premium && !user}
+                  >
+                    Read More
+                  </Button>
+                </Card.Body>
+              </Card>
+            </Col>
+          ))}
+        </Row>
+      )}
+    </Container>
   );
 };
 
